==> 1_Cleaned_4_415.ipynb <==
import numpy as np
import pandas as pd
import Cleanup

# define the function to be called later
cleanup = Cleanup.clean() #calling out the function 'Cleanup'

df = pd.read_csv("EMD_Reconstructed(i80 0400-0415).csv") #read csv file

df

df = cleanup.filter_non_zero_values(df) # to remove or filter out any rows that contain non-zero values

df

df = cleanup.filter_by_desired_lanes(df) #keep only the rows in the DataFrame that correspond to the specified lane(s)

df

df_filtered = cleanup.remove_lane_change(df) #remove or filter out rows from the original DataFrame

l = ['Vehicle_ID', 'Frame_ID', 'Total_Frames', 'v_Class', 'Lane_ID', 'Preceding', 'Following']
for col in l:
    df_filtered[col] = df_filtered[col].astype(int) #converts all the values to integers

### Adding two new column i80 and location

num_rows = df_filtered.shape[0]
df_filtered["Location"] = ['i80']*num_rows #populates all rows in that column with the value 'i80 on all rows
df_filtered["Period"] = ['4:00PM - 4:15PM']*num_rows #populates all rows in that column with the value '4:00PM - 4:15PM' on all rows

df_filtered.to_csv("Clean_4_415.csv") #saves the dataset into csv file named "Clean_4_415"
