==> 01.1_Merge_all_cleaned_files.ipynb <==
# import libaries
import pandas as pd 
import numpy as np 

# Read the data
df1 = pd.read_csv("Clean_4_415.csv")
df2 = pd.read_csv("Cleaned_5_515.csv")
df3 = pd.read_csv("Cleaned_515_530.csv")

### Combining all dataset with different time period

# Combine the datasets
combined_i80 = pd.concat([df1,df2,df3], axis=0)

# Reset the index of the combined dataset
combined_i80.reset_index(drop=True, inplace=True)
combined_i80

# Save the DataFrame 'combined_i80' to a CSV file named "Merged_i80.csv" without including the index column in the CSV file
combined_i80.to_csv("Merged_i80.csv",index = False)
