==> 03_Transformation_01.ipynb <==
# import libraries
import numpy as np 
import pandas as pd 
import Cleanup

# Call the clean() method from the Cleanup class to perform data cleanup or transformation
clean = Cleanup.clean()

# Read the data
df_i = pd.read_csv("Merged_i80.csv")


# Drop the 'Unnamed: 0' column from the DataFrame df_i
df_i = df_i.drop('Unnamed: 0', axis = 1)
df_i

# Convert the input variables from feet to meters in df_i DataFrame
df_i['v_Length'] = df_i['v_Length'] * 0.3048
df_i['Space_Headway'] = df_i['Space_Headway'] * 0.3048
df_i['v_Vel'] = df_i['v_Vel'] * 0.3048
df_i['v_Acc'] = df_i['v_Acc'] * 0.3048
df_i['Local_X'] = df_i['Local_X'] * 0.3048
df_i['Local_Y'] = df_i['Local_Y'] * 0.3048



## Merge DataFrame based on Preceding Vehicle ID and Frame ID
df_merged = df_i.merge(df_i, left_on=['Preceding', 'Frame_ID'], right_on=['Vehicle_ID', 'Frame_ID'], suffixes=('', '_preceding'))

df_merged

# Combine the 'Preceding' and 'Vehicle_ID' columns to create a new column 'L-F_Pair' in the DataFrame df_merged
df_merged['L-F_Pair'] = df_merged['Preceding'].astype(str) + \
            '-' + df_merged['Vehicle_ID'].astype(str)

# Call the 'time_pairs' method from the 'clean' object and assign the result to the DataFrame 'df_merged'
df_merged = clean.time_pairs(df_merged)

df_merged

# Write the DataFrame 'df_merged' to a CSV file named 'merged.csv' without including the index column
df_merged.to_csv('merged.csv',index = False)

# Extract a subset of the DataFrame 'df_merged' containing only the columns 'Vehicle_ID' and 'Preceding'
df_merged[['Vehicle_ID', 'Preceding']]
