==> 03_Transformation_02.ipynb <==
# import libraries
import numpy as np
import pandas as pd
import Cleanup

# Call the clean() method from the Cleanup class to perform data cleanup or transformation
clean = Cleanup.clean()

df = pd.read_csv("merged.csv")
df.head()

df.columns

# Drop the specified columns from the DataFrame 'df'
df = df.drop(['Preceding_preceding','Following_preceding','Local_X_preceding', 'Local_Y_preceding',
              'Global_X_preceding', 'Global_Y_preceding','Total_Frames_preceding', 'Location_preceding', 'Period_preceding'],axis=1)

# So , for proper knowing the Vehicle type we are assigning the v_class with Vehicle class
 #   1 L Motorcycle   2: Car     3: Heavy Vehicle
    
filtered_vClass = df[['Vehicle_ID', 'v_Class', 'v_Length']].drop_duplicates().sort_values('v_Class')

class_mapping = {1: 'Motorcycle', 2: 'Car', 3: 'Heavy Vehicle'}

df['Vehicle_Class'] = df['v_Class'].map(class_mapping)


### Calculating the space gap by subtracting the length of the preceding vehicle from the space headway using the following code

# Calculate the space gap 

df['Space_Gap'] = df['Space_Headway'] - df['v_Length_preceding']

# Define the reaction times in seconds

df['Difference_of_speed'] = df['v_Vel_preceding'] - df['v_Vel']



# Filter the dataset to remove vehicles with time headway >= 5 seconds
df = df.groupby(['Vehicle_ID', 'Period']).filter(lambda x: (x['Time_Headway'] < 5).all()).reset_index(drop=True)


df

df.to_csv("Transformation_2.csv", index = False)
