==> 3_Clean_515_530.ipynb <==
import pandas as pd 
import numpy as np 
import Cleanup

# define the function to be called later
clean = Cleanup.clean() #calling out the function 'Cleanup'

df = pd.read_csv("EMD_Reconstructed(i80 0515-0530).csv") #read csv file
df.head()

df = clean.filter_non_zero_values(df) #to remove or filter out any rows that contain non-zero values

df = clean.filter_by_desired_lanes(df) #keep only the rows in the DataFrame that correspond to the specified lane(s)


df

df_filtered= clean.remove_lane_change(df) #df_filtered will contain the DataFrame with lane changes removed

l = ['Vehicle_ID', 'Frame_ID', 'Total_Frames', 'v_Class', 'Lane_ID', 'Preceding', 'Following']
for col in l:
    df_filtered[col] = df_filtered[col].astype(int) #converts all the values to integers

df_filtered

### Adding two new column(LOcation and time period)

num_rows = df_filtered.shape[0]
df_filtered["Location"] = ['i80']*num_rows #populates all rows in that column with the value 'i80 on all rows
df_filtered["Period"] = ['5:15PM - 5:30PM']*num_rows #num_rows populates all rows in that column with the value '5:15PM - 5:30PM' on all rows

df_filtered.to_csv("Cleaned_515_530.csv") #saves the dataset into csv file named "Clean_515_530"
