==> 05_SVM_MODEL_Accerlation final.ipynb <==
# Import libraries
import numpy as np 
import pandas as pd 
from sklearn.svm import SVC
from sklearn.svm import SVR
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import LabelEncoder

import Cleanup
import warnings
# Ignore specific warning messages from a package
warnings.filterwarnings("ignore")


# Read the data
train_df = pd.read_csv("train_df.csv")
test_df = pd.read_csv("test_df.csv")


# Drops the column named 'Unnamed: 0' from two DataFrames, 'train_df' and 'test_df'
train_df = train_df.drop(['Unnamed: 0'],axis = 1)
test_df = test_df.drop(['Unnamed: 0'],axis = 1)

# Create a LabelEncoder object
label_encoder = LabelEncoder()
# Encode the 'L-F_Pair' column in 'train_df' and 'test_df' using the LabelEncoder and update the DataFrame
train_df['L-F_Pair'] = label_encoder.fit_transform(train_df['L-F_Pair'])
test_df['L-F_Pair'] = label_encoder.fit_transform(test_df['L-F_Pair'])

 def create_prediction_columns(df, n):
        '''
        create the prediction pair by shifting the actual data up by the mentioned number(0.1*n seconds) to create the timeseries info
        '''
        df["nextframeAcc"] = df.groupby(
            ["L-F_Pair"], as_index=False)["v_Acc"].shift(-1*n)
        df["nextframesvel"] = df.groupby(
            ["L-F_Pair"], as_index=False)["v_Vel"].shift(-1*n)
        df["nextframeposition"] = df.groupby(
            ["L-F_Pair"], as_index=False)["Local_Y"].shift(-1*n)
        df["nextFrameSpacing"] = df.groupby(
            ["L-F_Pair"], as_index=False)["Space_Gap"].shift(-1*n)
        df["precnextframesvel"] = df.groupby(
            ["L-F_Pair"], as_index=False)["v_Vel_preceding"].shift(-1*n)
        df = df[df['nextframeposition'].notna()]
        df = df[df['nextframesvel'].notna()]
        df = df[df['nextframeAcc'].notna()]
        df = df[df['nextFrameSpacing'].notna()]
        df = df[df['precnextframesvel'].notna()]

        return df

# Use the function 'create_prediction_columns' to create prediction pairs for the DataFrame 'train_df' with a shifting window of 5 time steps
train_df = create_prediction_columns(train_df,5)

#  Use the function 'create_prediction_columns' to create prediction pairs for the DataFrame 'test_df' with a shifting window of 5 time steps
test_df = create_prediction_columns(test_df,5)

train_df

### SUPPORT VECTOR MACHINE

# Prepare the feature matrix 'X_train' by dropping the columns 'nextframeAcc', 'Vehicle_Class', 'Location', and 'Period' from the DataFrame 'train_df'
X_train = train_df.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
# Prepare the target array 'y_train' containing the values from the 'nextframeAcc' column of the DataFrame 'train_df'
y_train = train_df['nextframeAcc']
X_test = test_df.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
y_test = test_df['nextframeAcc']

The code performs feature scaling using the StandardScaler to standardize the features in both the training and test datasets. Standardization is a preprocessing step that scales the features to have a mean of 0 and a standard deviation of 1, which is a common requirement for many machine learning algorithms.

# Create a StandardScaler object named 'sc'
sc = StandardScaler()
#  Fit the StandardScaler on the training feature matrix 'X_train' to compute the mean and standard deviation of the features
sc.fit(X_train)
# # Use the fitted StandardScaler to transform the training feature matrix 'X_train' and 'X_test'to standardized values, and store the result in 'X_train_std' and 'X_test'
X_train_std = sc.transform(X_train)
X_test_std = sc.transform(X_test)

# Instantiate the Support Vector Regressor 
svr= SVR(C=1.0, epsilon=0.2)
 
# Fit the model
svr.fit(X_train_std, y_train)

The code defines a function named 'predict_velocity_and_spacing', which takes a DataFrame 'df' as input and returns a new DataFrame containing the predicted velocity and spacing values for each time step in the original DataFrame. The function uses a shifting window approach to predict the velocity and spacing for each time step based on previous time step values.

def predict_velocity_and_spacing(df, target_variable='nextframeAcc', delta_time = 0.1,model=svr):
    predicted_df = []

    for current_pair in df['L-F_Pair'].unique():
        input_df = df[df['L-F_Pair'] == current_pair].copy()
# Initialize arrays to store velocity, predicted acceleration, spacing, and local_y for each time step 
        vel = np.zeros(input_df.shape[0])
        pred_acc = np.zeros(input_df.shape[0])
        spacing = np.zeros(input_df.shape[0])
        local_y_subject = np.zeros(input_df.shape[0])
        local_y_preceding = np.zeros(input_df.shape[0])
  # Initialize the first time step values from the input DataFrame
        vel[0] = input_df.iloc[0]['v_Vel']
        spacing[0] = input_df.iloc[0]['Space_Gap']
        pred_acc[0] = input_df.iloc[0][target_variable]
        local_y_subject[0] = input_df.iloc[0]['Local_Y']
        local_y_preceding[0] = input_df.iloc[0]['Local_Y_preceding']

        for j in range(1, len(input_df)):
            # Calculate the velocity for the current time step using the previous velocity and predicted acceleration
            vel[j] = vel[j - 1] + (pred_acc[j - 1] * delta_time)
            if vel[j] < 0:
                vel[j] = 0
            spacing[j] = spacing[j - 1] + (vel[j - 1] * delta_time) + (0.5 * pred_acc[j - 1] * pow(delta_time, 2))
            local_y_subject[j] = local_y_subject[j - 1] + (vel[j - 1] * delta_time) + (0.5 * pred_acc[j - 1] * pow(delta_time, 2))
            local_y_preceding[j] = input_df.iloc[j]['Local_Y_preceding']

            # Calculate the predicted acceleration for the next time step
            #pred_acc[j] = model.predict(np.array([spacing[j], input_df.iloc[j]['v_Class_preceding'], input_df.iloc[j]['v_Class'], vel[j], vel[j] - input_df.iloc[j]['v_Vel_preceding']]).reshape(1, -1))

        input_df['predicted_velocity'] = vel
        input_df['predicted_spacing'] = spacing
        predicted_df.append(input_df)

    result = pd.concat(predicted_df)
    return result


# Use the trained SVR model 'svr' to predict the target variable for the test feature matrix 'X_test'
pred_5 = svr.predict(X_test)

# Assign the predicted values ('pred_5') to the 'nextframeAcc' column of the test feature matrix 'X_test'
X_test["nextframeAcc"]=pred_5

# Assuming you have trained the Random Forest model (rf) and X_test is the test DataFrame
predictions_svm_5= predict_velocity_and_spacing(X_test, target_variable='nextframeAcc')
predictions_svm_5

### Calculating R2 and RMSE 

from sklearn.metrics import r2_score, mean_absolute_error,mean_squared_error

r2_score(y_test, pred_5) #evaluating the model's performance against the ground truth on the test set

mse = mean_squared_error(y_test,pred_5)
rmse = np.sqrt(mse)

print("RMSE:", rmse)

### Reaction time of 1s

# Create a new DataFrame 'train_df_1' with additional columns representing time-series information based on a shifting window of 10 time steps
train_df_1 = create_prediction_columns(train_df,10)

# Create a new DataFrame 'test_df_1' with additional columns representing time-series information based on a shifting window of 10 time steps
test_df_1 = create_prediction_columns(test_df,10)

# Prepare the training and test datasets by separating features and target variables
# For the training dataset:
X_train = train_df_1.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
y_train = train_df_1['nextframeAcc']
# For the test dataset:
X_test = test_df_1.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
y_test = test_df_1['nextframeAcc']

from sklearn.preprocessing import StandardScaler
from sklearn.svm import SVR
sc = StandardScaler()
# Fit the StandardScaler on the training features to learn the mean and standard deviation of each feature
sc.fit(X_train)
# Transform the training features using the learned mean and standard deviation
X_train_std = sc.transform(X_train)
# Transform the test features using the same mean and standard deviation as learned from the training data
X_test_std = sc.transform(X_test)

# Instantiate the Support Vector Regressor 
svr= SVR(C=1.0, epsilon=0.2)
 
# Fit the model
svr.fit(X_train_std, y_train)

# Use the trained SVR model to predict the target variable ('nextframeAcc') for the test features ('X_test')
pred_1 = svr.predict(X_test)
# Update the 'nextframeAcc' column in the test feature matrix ('X_test') with the predicted values
X_test["nextframeAcc"]=pred_1

# Assuming you have trained the Random Forest model (rf) and X_test is the test DataFrame
predictions_svm_1 = predict_velocity_and_spacing(X_test, target_variable='nextframeAcc')
predictions_svm_1

### Calculating R2 and RMSE 

r2_score(y_test, pred_1) #evaluating the model's performance against the ground truth on the test set

mse = mean_squared_error(y_test,pred_1)
rmse = np.sqrt(mse)

print("RMSE:", rmse)

### Reaction Time 1.5 s

# Create a new DataFrame 'train_df_15' with additional columns representing time-series information based on a shifting window of 15 time steps
train_df_15 = create_prediction_columns(train_df,15)

# Create a new DataFrame 'test_df_15' with additional columns representing time-series information based on a shifting window of 15 time steps
test_df_15 = create_prediction_columns(test_df,15)

# Prepare the training and test datasets by separating features and target variables
# For the training dataset:
X_train = train_df_15.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
y_train = train_df_15['nextframeAcc']
# For the test dataset:
X_test = test_df_15.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
y_test = test_df_15['nextframeAcc']

sc = StandardScaler()
# Fit the StandardScaler on the training features to learn the mean and standard deviation of each feature
sc.fit(X_train)
# Transform the training features using the learned mean and standard deviation
X_train_std = sc.transform(X_train)
# Transform the test features using the same mean and standard deviation as learned from the training data
X_test_std = sc.transform(X_test)

# Instantiate the Support Vector Regressor 
svr= SVR(C=1.0, epsilon=0.2)
 
# Fit the model
svr.fit(X_train_std, y_train)

# Use the trained Support Vector Regressor (SVR) model to predict the target variable ('nextframeAcc') for the test features ('X_test')
pred_15 = svr.predict(X_test)
# Update the 'nextframeAcc' column in the test feature matrix ('X_test') with the predicted values
X_test["nextframeAcc"]=pred_15

# Assuming you have trained the Random Forest model (rf) and X_test is the test DataFrame
predictions_svm_15 = predict_velocity_and_spacing(X_test, target_variable='nextframeAcc')
predictions_svm_15

### Calculating R2 and RMSE 

r2_score(y_test, pred_15) #evaluating the model's performance against the ground truth on the test set

mse = mean_squared_error(y_test,pred_15)
rmse = np.sqrt(mse)

print("RMSE:", rmse)

### Reaction Time of 2s

# Create a new DataFrame 'train_df_2' with additional columns representing time-series information based on a shifting window of 20 time steps
train_df_2= create_prediction_columns(train_df,20)

# Create a new DataFrame 'test_df_2' with additional columns representing time-series information based on a shifting window of 20 time steps
test_df_2= create_prediction_columns(test_df,20)

# Prepare the training and test datasets by separating features and target variables
# For the training dataset:
X_train = train_df_2.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
y_train = train_df_2['nextframeAcc']
# For the test dataset:
X_test = test_df_2.drop(['nextframeAcc', 'Vehicle_Class', 'Location', 'Period'], axis = 1)
y_test = test_df_2['nextframeAcc']

sc = StandardScaler()
# Fit the StandardScaler on the training features to learn the mean and standard deviation of each feature
sc.fit(X_train)
# Transform the training features using the learned mean and standard deviation
X_train_std = sc.transform(X_train)
# Transform the test features using the same mean and standard deviation as learned from the training data
X_test_std = sc.transform(X_test)

# Instantiate the Support Vector Regressor 
svr= SVR(C=1.0, epsilon=0.2)
 
# Fit the model
svr.fit(X_train_std, y_train)

# Use the trained Support Vector Regressor (SVR) model to predict the target variable ('nextframeAcc') for the test features ('X_test')
pred_2 = svr.predict(X_test)
# Update the 'nextframeAcc' column in the test feature matrix ('X_test') with the predicted values
X_test["nextframeAcc"]=pred_2

# Assuming you have trained the Random Forest model (rf) and X_test is the test DataFrame
predictions_svm_2 = predict_velocity_and_spacing(X_test, target_variable='nextframeAcc')
predictions_svm_2

### Calculating R2 and RMSE 

r2_score(y_test, pred_2) #evaluating the model's performance against the ground truth on the test set

mse = mean_squared_error(y_test,pred_2)
rmse = np.sqrt(mse)

print("RMSE:", rmse)
